test3

+++ server/src/test/java/ru/practicum/shareit/user/UserServiceTest.java
@@ -27,7 +27,7 @@ public class UserServiceTest {
         UserAddDtoOut userAddDtoOut = userService.addUser(userAddDtoIn);
         Assertions.assertEquals(userAddDtoOut.getEmail(), userAddDtoIn.getEmail());
         Assertions.assertEquals(userAddDtoOut.getName(), userAddDtoIn.getName());
-        Assertions.assertEquals(userAddDtoOut.getId(), (long)1);
+        Assertions.assertEquals(userAddDtoOut.getId(), 1);
     }
 
     @Test
